% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admix_props_amr}
\alias{admix_props_amr}
\title{Admixture proportions of AMR subset of 1000 Genomes dataset}
\format{
A 3 \eqn{\times} 353 matrix of admixture proportions. Each column is the admixture proportions of the
corresponding individuals.
}
\usage{
admix_props_amr
}
\description{
Admixture proportions of AMR subset of 1000 Genomes dataset
}
\keyword{datasets}
