% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admix_props_hgdp}
\alias{admix_props_hgdp}
\title{Admixture proportions of HGDP dataset}
\format{
A 3 \eqn{\times} 929 matrix of admixture proportions. Each column is the admixture proportions of the
corresponding individuals.
}
\usage{
admix_props_hgdp
}
\description{
Admixture proportions of HGDP dataset
}
\keyword{datasets}
