% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coanc_pops_amr}
\alias{coanc_pops_amr}
\title{Coancestry among populations of AMR subset of 1000 Genomes dataset}
\format{
A 3 \eqn{\times} 3 matrix of coancestry among populations. Each entry is coancestry coefficient for
the corresponding pair of the antecedent populations.
}
\usage{
coanc_pops_amr
}
\description{
Coancestry among populations of AMR subset of 1000 Genomes dataset
}
\keyword{datasets}
