% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coanc_pops_hgdp}
\alias{coanc_pops_hgdp}
\title{Coancestry among populations of HGDP dataset}
\format{
A 7 \eqn{\times} 7 matrix of coancestry among populations. Each entry is coancestry coefficient for
the corresponding pair of the antecedent populations.
}
\usage{
coanc_pops_hgdp
}
\description{
Coancestry among populations of HGDP dataset
}
\keyword{datasets}
