% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_indep_antepops.R
\name{compute_null}
\alias{compute_null}
\title{Compute a null test statistics for testing of no coancestry among antecedent populations}
\usage{
compute_null(p_anc, admix_proportions, inbr_antepops, verbose = FALSE, ...)
}
\arguments{
\item{p_anc}{A length-\eqn{m} vector representing allele frequencies at the ancestral population.}

\item{admix_proportions}{A \eqn{K} \eqn{\times} \eqn{n} matrix representing admixture proportions.}

\item{inbr_antepops}{A length-\eqn{K} vector of inbreeding coefficient.}

\item{verbose}{A Boolean that controls the returned values. If \code{verbose} is FALSE (default), only the estimated \eqn{\boldsymbol{\Lambda}}
is returned. If TRUE, \code{est_coanc} returns a list consisting of the estimated \eqn{\boldsymbol{\Lambda}},
a list of \eqn{\boldsymbol{\Lambda}}s over iterations, a list of errors over iterations.}

\item{...}{Other options used to control the estimation of coancestry in \code{\link[popkin:popkin]{popkin::popkin()}}. Passed on to \code{\link[popkin:popkin]{popkin::popkin()}}.}
}
\value{
A test statistic under the null.
}
\description{
\code{compute_null()} computes a null test statistic for testing of no coancestry among antecedent populations.
}
\seealso{
\code{\link[=test_indep_antepops]{test_indep_antepops()}} for hypothesis testing of no coancestry among antecedent populations.
}
\keyword{internal}
