% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbl_admixture.R
\name{dbl_admixture}
\alias{dbl_admixture}
\title{Simulating allele frequencies and genotypes by the double-admixture method}
\usage{
dbl_admixture(
  p_anc,
  coanc_antepops,
  admix_proportions,
  paras = NULL,
  paras_only = FALSE,
  p_antepops_only = FALSE,
  geno_only = FALSE,
  ...
)
}
\arguments{
\item{p_anc}{A length-\eqn{m} vector representing allele frequencies at the ancestral population.}

\item{coanc_antepops}{A \eqn{K \times K} matrix representing coancestry among antecedent populations.}

\item{admix_proportions}{A \eqn{K} \eqn{\times} \eqn{n} matrix representing admixture proportions.}

\item{paras}{An optional named list containing \code{Gamma} and \code{W}. If provided, \code{Gamma} and \code{W} won't be estimated. See details in \link{est_paras_dbladmix}.}

\item{paras_only, p_antepops_only, geno_only}{Booleans that control the returned values. If \code{paras_only = T}, the estimated \eqn{\boldsymbol{W}}
and \eqn{\boldsymbol{\Gamma}} are returned. If \code{p_antepops_only = T}, only the \eqn{m \times K} matrix of antecedent allele frequencies
\eqn{\boldsymbol{P}} is returned. If \code{geno_only = T}, only the \eqn{m \times n} matrix of genotypes \eqn{\boldsymbol{X}} is returned. Otherwise,
a list containing \eqn{\boldsymbol{W}}, \eqn{\boldsymbol{\Gamma}}, \eqn{\boldsymbol{P}} and \eqn{\boldsymbol{X}} is returned.}

\item{...}{Other options used to control parameter estimation. Passed on to \link{est_paras_dbladmix}.}
}
\value{
\itemize{
\item If \code{paras_only = T}, returns parameters estimated from \link{est_paras_dbladmix}.
\item If \code{p_antepops = T}, an \eqn{m \times K} matrix of allele frequencies of antecedent populations \eqn{\boldsymbol{P}} is returned.
\item If \code{geno_only = T}, an \eqn{m \times n} matrix of genotypes is returned.
\item Otherwise, a named list is returned, containing:
\itemize{
\item \code{W} and \code{Gamma}: the parameters estimated from \link{est_paras_dbladmix},
\item \code{p_antepops}: the simulated \eqn{m \times K} matrix of antecedent allele frequencies,
\item \code{X}: the simulated \eqn{m \times n} matrix of genotypes.
}
}
}
\description{
\code{dbl_admixture()} generates random samples of antecedent allele frequencies \eqn{\boldsymbol{P}} and random samples of genotypes \eqn{\boldsymbol{X}}
according to the double-admixture method.
}
\examples{
## Generate a genotype from HGDP -------------------------------------------------------------------
## It takes 87.842 sec to finish the following codes on a 2.3 GHz 18-Core Intel Xeon W iMac Pro.
\dontrun{
data("fam_hgdp",         package = "superadmixture")
data("p_anc_hgdp",       package = "superadmixture")
data("admix_props_hgdp", package = "superadmixture")
data("coanc_pops_hgdp",  package = "superadmixture")
X_hgdp <- dbl_admixture(
      p_anc = p_anc_hgdp,
      coanc_antepops = coanc_pops_hgdp,
      admix_proportions = admix_props_hgdp,
      geno_only = TRUE)

## Estimate kinship from the simulated genotypes
kinship <- popkin::popkin(X_hgdp)
coanc_indiv <- popkin::inbr_diag(kinship)
coanc_indiv <- ifelse(coanc_indiv < 0, 0, coanc_indiv)

## Visualize kinship
popkin::plot_popkin(coanc_indiv, labs = fam_hgdp$subpop, labs_las = 2)}
}
\seealso{
\link{est_paras_dbladmix} for details about estimating parameters of the Double-Admixture simulation.
}
