% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_p_indiv.R
\name{est_p_indiv}
\alias{est_p_indiv}
\title{Estimate Individual specific Allele Frequencies}
\usage{
est_p_indiv(
  X,
  k_antepops,
  loci_on_cols = FALSE,
  rowspace_only = FALSE,
  p_indiv_only = FALSE
)
}
\arguments{
\item{X}{A matrix of genotypes. \code{X} should be an integer matrix of 0's, 1's, 2's or \code{NA}.}

\item{k_antepops}{An integer representing the number of antecedent populations.}

\item{loci_on_cols}{A boolean that specifies the dimension of the input \code{X}. If \code{FALSE} (default), \code{X} has \eqn{m} (number of loci) \eqn{\times} \eqn{n} (number of individuals) dimension.
If \code{TRUE}, \code{X}'s dimension is \eqn{n \times m}.}

\item{rowspace_only, p_indiv_only}{Booleans for controlling the return values. If \code{rowspace_only = T}, an \eqn{n \times K} matrix of eigen-vectors \eqn{\boldsymbol{W}}
of \eqn{\boldsymbol{G}} is returned. If \code{p_indiv_only = T}, an \eqn{m \times n} matrix \eqn{\boldsymbol{\Pi}} of estimated individual specific allele frequencies is returned.
Otherwise, a list that containing \eqn{\boldsymbol{W}} and \eqn{\boldsymbol{\Pi}} is returned.}
}
\value{
\itemize{
\item If \code{rowspace_only = T}, an \eqn{n \times K} matrix of eigen-vectors \eqn{\boldsymbol{W}} of \eqn{\boldsymbol{G}} is returned.
\item If \code{p_indiv_only = T},  an \eqn{m \times n} matrix \eqn{\boldsymbol{\Pi}} of estimated individual specific allele frequencies is returned.
\item Otherwise, a named list is returned, containing:
\itemize{
\item \code{p_indiv}: The \eqn{m \times n} matrix of estimated IAFs,
\item \code{rowspace}: The \eqn{n \times K} matrix of eigen-vectors \eqn{\boldsymbol{W}} of \eqn{\boldsymbol{G}}.
}
}
}
\description{
\code{est_p_indiv()} estimates the individual specific allele frequencies \eqn{\boldsymbol{\Pi}} from a genotype matrix \eqn{\boldsymbol{X}}.
}
\details{
\code{est_p_indiv()} adopts a three-step strategy to estimate the individual specific allele frequencies.
\enumerate{
\item If \code{loci_on_cols}, \eqn{\boldsymbol{X} \leftarrow t(\boldsymbol{X})}.
\item Calculates the top \eqn{K} eigen-vectors \eqn{\boldsymbol{W}} of the matrix \eqn{\boldsymbol{G} = \frac{1}{m} \boldsymbol{X}^{\prime}\boldsymbol{X} - \boldsymbol{D}}
where the matrix \eqn{\boldsymbol{D}} is a diagonal matrix used to adjust for heteroskedasticity for the Binomial distribution. More specifically,
\eqn{d_{jj}} is defined as \deqn{d_{jj} = \frac{1}{m}\sum_{i = 1}^m (2 x_{ij} - x_{ij}^2)} where \eqn{x_{ij}} represents the genotypic value for
the individual \eqn{j} at the \eqn{i}th locus.
\item Estimate the individual specific allele frequencies \eqn{\boldsymbol{\Pi}}  by projecting \eqn{\boldsymbol{X}} onto \eqn{\boldsymbol{W}}:
\eqn{\boldsymbol{\Pi} = \frac{1}{2} \boldsymbol{X}\boldsymbol{W}\boldsymbol{W}^{\prime}}. If \eqn{\pi_{ij} > 1} or \eqn{\pi_{ij} < 0}, \eqn{\pi_{ij}} is set to be
\eqn{\min(1, \max(0, \pi_{ij}))}.
}
}
\examples{
## Estimate Individual Specific Allele Frequencies of HGDP  -----------------------------
## It takes 6.451 sec to finish the following codes on a 2.3 GHz 18-Core Intel Xeon W iMac Pro.
data("X_hgdp", package = "superadmixture")
p_indiv <- est_p_indiv(X_hgdp, k_antepops = 7, loci_on_cols = TRUE, p_indiv_only = TRUE)
}
