% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norta.R
\name{estimate_corrmat}
\alias{estimate_corrmat}
\title{Estimate the correlation matrix \eqn{\boldsymbol{\Sigma}_{\boldsymbol{z}}} used in the NORTA method}
\usage{
estimate_corrmat(
  p_anc,
  coanc_antepops,
  max_iters = 100,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{p_anc}{A scalar between 0 and 1.}

\item{coanc_antepops}{A \eqn{K \times K} matrix representing coancestry among antecedent populations.}

\item{max_iters}{Maximum number of iterations.}

\item{tol}{A scalar for stopping criteria. If \eqn{|g(\rho)|} is less than \code{tol}, the Newton's method halts.}
}
\value{
A \eqn{K \times K} matrix \eqn{\boldsymbol{\Sigma}_{\boldsymbol{z}}} of the correlation matrix standard multivariate Normal distribution that will be used by \code{\link[=norta]{norta()}}.
}
\description{
\code{estimate_corrmat()} adopts the Newton-Raphson method to estimate the the correlation matrix \eqn{\boldsymbol{\Sigma}_{\boldsymbol{z}}} used in \link{norta}.
}
\details{
This function finds the \eqn{\boldsymbol{\Sigma}_{\boldsymbol{z}}} used in \link{norta}. This problem can be reduced to \eqn{K \times (K - 1) / 2} independent problems:
for each \eqn{u < v}, we need to find such \eqn{\rho} that
\deqn{\mathbb{E}[X_u X_v] = a^2 + a(1 - a)\lambda_{uv}}
We note that there is no close form soluation for \eqn{\rho}. We adopt the numeric integration to perform a numerical search for \eqn{\rho}.
Define
\deqn{g(\rho) = \mathbb{E}[F_u^{-1}(\Phi(z_u))F_v^{-1}(\Phi(z_v))] - a^2 - a(1 - a)\lambda_{uv}}
\deqn{g^{\prime}(\rho) = \mathbb{E}[F_u^{-1}(\Phi(z_u))F_v^{-1}(\Phi(z_v))\left(\frac{\rho}{1 - \rho^2} + \frac{z_u z_v - \rho(z_u^2 - \rho z_u z_v + z_v^2)}{(1 - \rho^2)^2}\right)]}
where \eqn{\Phi} denotes the univariate standard normal cumulative density function (cdf); \eqn{F_u} denotes the Balding-Nichols(\eqn{a_i}, \eqn{\lambda_{uu}});
\eqn{F_u^{-1}} denotes the inverse cdf of \eqn{F_u}.The iteration of Newton's method is
\deqn{\rho \leftarrow \rho \leftarrow \frac{g(\rho)}{g^{\prime}(\rho)}}
}
\seealso{
\link{norta} for details about simulating allele frequencies and genotypes.
}
\keyword{internal}
