% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_p_indiv.R
\name{factor_p_indiv}
\alias{factor_p_indiv}
\title{Factor Individual specific Allele Frequencies}
\usage{
factor_p_indiv(
  p_indiv,
  k_antepops,
  init_method = "rowspace",
  rowspace = NULL,
  loci_on_cols = FALSE,
  max_iters = 1e+06,
  tol = 1e-04,
  verbose = FALSE
)
}
\arguments{
\item{p_indiv}{A matrix of individual specific allele frequencies \eqn{\boldsymbol{\Pi}}. Each \eqn{\pi_{ij}} should be an numeric matrix each element lies between 0 and 1.}

\item{k_antepops}{An integer representing the number of antecedent populations.}

\item{init_method}{One of "random" or "rowspace" (default). If \code{init_method} is "random",
the columns of \eqn{\boldsymbol{Q}} are drawn from Dirichlet distribution. If \code{init_method} is "rowspace",
\eqn{\boldsymbol{Q}} are initialized by projecting \code{rowspace} to the simplex.}

\item{rowspace}{An optional \eqn{n \times K} dimensional rowspace of \eqn{\boldsymbol{Q}}.}

\item{loci_on_cols}{A boolean that specifies the dimension of the input \code{X}. If \code{FALSE} (default), \code{X} has \eqn{m} (number of loci) \eqn{\times} \eqn{n} (number of individuals) dimension.
If \code{TRUE}, \code{X}'s dimension is \eqn{n \times m}.}

\item{max_iters}{Maximum number of iterations.}

\item{tol}{A scalar for convergence criterion. If the relative change of \eqn{\|\boldsymbol{\Pi} - \boldsymbol{P}\boldsymbol{Q}\|_F} is less than \code{tol},
the algorithm halts.}

\item{verbose}{A Boolean that controls the returned values. If \code{verbose = F} (default), only the estimated \eqn{\boldsymbol{P}} and \eqn{\boldsymbol{Q}}
are returned. If \code{T}, this function returns a list consisting of the estimated \eqn{\boldsymbol{P}},
the estimated \eqn{\boldsymbol{Q}} and a list of errors over iterations.}
}
\value{
\itemize{
\item If \code{verbose = T}, a named list is returned, containing:
\itemize{
\item \code{P_hat}: The \eqn{m \times K} matrix of estimated \eqn{\boldsymbol{P}},
\item \code{Q_hat}: The \eqn{K \times n} matrix of estimated \eqn{\boldsymbol{Q}},
\item \code{err_list}: A list of \eqn{\|\boldsymbol{\Pi} - \boldsymbol{P}\boldsymbol{Q}\|_F}'s over iterations.
}
\item Otherwise, a named list containing \code{P_hat} and \code{Q_hat} is returned.
}
}
\description{
\code{factor_p_indiv()} factors the individual specific allele frequencies \eqn{\boldsymbol{\Pi}} into two matrices \eqn{\boldsymbol{P}} and \eqn{\boldsymbol{Q}} such that
\eqn{\boldsymbol{\Pi} = \boldsymbol{P}\boldsymbol{Q}}.
}
\details{
\code{factor_p_indiv()} finds matrices \eqn{\boldsymbol{P}} and \eqn{\boldsymbol{Q}} such that \eqn{\boldsymbol{\Pi} = \boldsymbol{P}\boldsymbol{Q}}, by solving the following
optimization problem
\deqn{\arg\min_{\boldsymbol{P}, \boldsymbol{Q}} \|\boldsymbol{\Pi} - \boldsymbol{P}\boldsymbol{Q}\|_F}
where  \eqn{\boldsymbol{P}} and \eqn{\boldsymbol{Q}} obeys these conditions:
\enumerate{
\item \eqn{p_{iu} \in [0,1]  ~ \forall (i,u)}
\item \eqn{q_{uj} \ge 0  ~ \forall (u,j)} and \eqn{\sum_u q_{uj} = 1 ~ \forall j}
}
}
\examples{
## Estimate Admixture Proportions of HGDP  ----------------------------------------
## It takes 14.469 sec to finish the following codes on a 2.3 GHz 18-Core Intel Xeon W iMac Pro.
data("X_hgdp", package = "superadmixture")
obj <- est_p_indiv(X_hgdp, k_antepops = 7, loci_on_cols = TRUE)
rowspace <- obj$rowspace
p_indiv <- obj$p_indiv
admix_props <- factor_p_indiv(p_indiv, k_antepops = 7,
          rowspace = rowspace, tol = 1e-2, max_iters = 200)$Q_hat
}
