% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fam_amr}
\alias{fam_amr}
\title{PLINK FAM file of AMR subset of 1000 Genomes dataset}
\format{
A dataframe of PLINK FAM file of AMR subset of 1000 Genomes dataset.
\describe{
\item{fam}{The population code. MXL	abbreviates for Mexican Ancestry from Los Angeles USA.
PUR	abbreviates for Puerto Rican from Puerto Rica. CLM abbreviates for Colombian from Medellian, Colombia.
PEL abbreviates for	Peruvian from Lima, Peru.}
\item{id}{The individual ID of 1000 Genomes Project.}
\item{pat}{The father's ID of 1000 Genomes Project.}
\item{mat}{The mother's ID of 1000 Genomes Project.}
\item{sex}{The gender code. "1" denotes males. "2" denotes female. }
\item{pheno}{The case-control status. All are -9.}
}
}
\usage{
fam_amr
}
\description{
PLINK FAM file of AMR subset of 1000 Genomes dataset
}
\keyword{datasets}
