% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fam_hgdp}
\alias{fam_hgdp}
\title{PLINK FAM file of HGDP dataset}
\format{
A dataframe of PLINK FAM file of HGDP dataset.
\describe{
\item{fam}{The population code.}
\item{id}{The individual ID of HGDP.}
\item{pat}{The father's ID of HGDP.}
\item{mat}{The mother's ID of HGDP.}
\item{sex}{The gender code. "1" denotes males. "2" denotes female. }
\item{pheno}{The phenotypic value}
\item{population}{The population code.}
\item{latitude}{}
\item{longitude}{}
\item{subpop}{The super-population label.}
}
}
\usage{
fam_hgdp
}
\description{
PLINK FAM file of HGDP dataset
}
\keyword{datasets}
