% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tree.R
\name{get_current_node}
\alias{get_current_node}
\title{Return the node index of the current tree node}
\usage{
get_current_node(query, tr)
}
\arguments{
\item{query}{A string label of the current tree node.}

\item{tr}{A phylo object from the \code{ape} package.}
}
\value{
A scalar that corresponds to node index of the current tree node.
}
\description{
Return the node index of the current tree node
}
\keyword{internal}
