% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seq_colors.R
\name{get_seq_colors}
\alias{get_seq_colors}
\title{Get a sequence of colors from RColorBrewer palettes}
\usage{
get_seq_colors(palette, n_colors)
}
\arguments{
\item{palette}{A string of RColorBrewer palette name.}

\item{n_colors}{An integer of number of colors. \code{n_colors} must be greater than 0.}
}
\value{
A vector of Hex color codes.
}
\description{
\code{get_seq_colors()} returns a sequence of Hex color codes from RColorBrewer package.
}
\note{
\code{palette} must be one of the following: BrBG, PiYG, PRGn, PuOr, RdBu, RdGy, RdYlBu, RdYlGn, Spectral, Accent, Dark2, Paired, Pastel1, Pastel2, Set1, Set2, Set3, Blues, BuGn, BuPu, GnBu, Greens, Greys, Oranges, OrRd, PuBu, PuBuGn, PuRd, Purples, RdPu, Reds, YlGn, YlGnBu, YlOrBr, YlOrRd.
}
\examples{
## Getting three colors from 'Reds' palette --------------------------
get_seq_colors("Reds", 3)

## Getting 20 colors from 'Spectral' palette -------------------------
get_seq_colors("Spectral", 20)

}
\keyword{internal}
