% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_coanc_antepops.R
\name{heatmap_coanc_antepops}
\alias{heatmap_coanc_antepops}
\title{Visualize the coancestry matrix between antecedent populations as a heatmap}
\usage{
heatmap_coanc_antepops(
  coanc_antepops,
  tip_label = "S",
  tl.cex = 1.3,
  cl.cex = 1,
  tl.offset = 0.6
)
}
\arguments{
\item{coanc_antepops}{A \eqn{K \times K} matrix representing coancestry among antecedent populations.}

\item{tip_label}{An optional string of prefix of the tip label (default to "S").}

\item{tl.cex}{Numeric, for the size of text label (variable names).}

\item{cl.cex}{Numeric, cex of number-label in color-legend.}

\item{tl.offset}{Numeric, for text label, see \code{\link[corrplot]{corrplot}}.}
}
\value{
A heatmap of \eqn{\boldsymbol{\Lambda}}.
}
\description{
\code{heatmap_coanc_antepops()} serves as a wrapper function for the \code{\link[corrplot]{corrplot}} function from the corrplot package to plot the coancestry between antecedent population \eqn{\boldsymbol{\Lambda}} as a heatmap.
}
\examples{
## Visualizing the coancestry of HGDP data ----------------------------------------
data("coanc_pops_hgdp", package = "superadmixture")

## order in the ascending order of coancestry
index <- order(diag(coanc_pops_hgdp))
coanc_pops_hgdp <- coanc_pops_hgdp[index, index]

## label antecedent populations
colnames(coanc_pops_hgdp) <- rownames(coanc_pops_hgdp) <- paste0("S", 1:7)

## plot the heatmap of coancestry
par(xpd = TRUE)
heatmap_coanc_antepops(coanc_pops_hgdp)

}
