% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_anc_amr}
\alias{p_anc_amr}
\title{Ancestral allele frequencies of AMR subset of 1000 Genomes dataset}
\format{
A length 10,000 vector of allele frequencies.
}
\usage{
p_anc_amr
}
\description{
Ancestral allele frequencies of AMR subset of 1000 Genomes dataset
}
\keyword{datasets}
