% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_anc_hgdp}
\alias{p_anc_hgdp}
\title{Ancestral allele frequencies of HGDP dataset}
\format{
A length 10,000 vector of allele frequencies.
}
\usage{
p_anc_hgdp
}
\description{
Ancestral allele frequencies of HGDP dataset
}
\keyword{datasets}
