% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Sample from Dirichlet distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{An integer of number of samples.}

\item{alpha}{A length-\eqn{m} vector of shape parameters. \code{alpha} should be positive.}
}
\value{
An \eqn{n \times m} matrix of random samples from the Dirichlet(\eqn{\boldsymbol{\alpha}}) distribution.
}
\description{
\code{rdirichlet()} returns an \eqn{n \times m} matrix of random samples from the Dirichlet(\eqn{\boldsymbol{\alpha}}) distribution.
}
\keyword{internal}
