% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sHWE.R
\name{sHWE}
\alias{sHWE}
\title{Hardy-Weinberg Equilibrium in structure populations}
\usage{
sHWE(X, k_antepops, rowspace = NULL, loci_on_cols = FALSE)
}
\arguments{
\item{X}{A matrix of genotypes. \code{X} should be an integer matrix of 0's, 1's, 2's or \code{NA}.}

\item{k_antepops}{An integer representing the number of antecedent populations.}

\item{rowspace}{An optimal \eqn{n \times K} matrix of eigen-vectors \eqn{\boldsymbol{W}}. See details in \code{\link[=est_p_indiv]{est_p_indiv()}}.}

\item{loci_on_cols}{A boolean that specifies the dimension of the input \code{X}. If \code{FALSE} (default), \code{X} has \eqn{m} (number of loci) \eqn{\times} \eqn{n} (number of individuals) dimension.
If \code{TRUE}, \code{X}'s dimension is \eqn{n \times m}.}
}
\value{
A vector of p-values for each SNP.
}
\description{
Compute structured Hardy-Weinberg Equilibrium (sHWE) p-values on a SNP-by-SNP basis.
}
\examples{
# Performing sHWE test on AMR subset of TGP for K = 3 -------------------------------------
data("X_amr", package = "superadmixture")
rowspace <- est_p_indiv(X_amr, k_antepops = 3, loci_on_cols = TRUE, rowspace_only = TRUE)
p_values <- sHWE(X_amr, 3, rowspace, TRUE)
hist(p_values)
}
