% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superadmixture-package.R
\docType{package}
\name{superadmixture-package}
\alias{superadmixture}
\alias{superadmixture-package}
\title{superadmixture: Inference of Population Structure by Super-admixture Framework}
\description{
This package provides functions for inference of population coancestry and simulating genotypes according to the super-admixture framework. It also has functions for conducting the structured Hardy-Weinberg equilibrium test and estimating admixture proportions using the ALStructure algorithm.
}
\author{
\strong{Maintainer}: Danfeng Chen \email{danfengc@princeton.edu} (\href{https://orcid.org/0000-0002-4489-0485}{ORCID})

Authors:
\itemize{
  \item John D. Storey \email{jstorey@princeton.edu} (\href{https://orcid.org/0000-0001-5992-402X}{ORCID})
}

}
\keyword{internal}
