% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_indep_antepops.R
\name{test_indep_antepops}
\alias{test_indep_antepops}
\title{Test of no coancestry among antecedent populations}
\usage{
test_indep_antepops(
  X,
  admix_proportions,
  loci_on_cols = TRUE,
  B = 100,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{A matrix of genotypes. \code{X} should be an integer matrix of 0's, 1's, 2's or \code{NA}.}

\item{admix_proportions}{A \eqn{K} \eqn{\times} \eqn{n} matrix representing admixture proportions.}

\item{loci_on_cols}{A boolean that specifies the dimension of the input \code{X}. If \code{FALSE} (default), \code{X} has \eqn{m} (number of loci) \eqn{\times} \eqn{n} (number of individuals) dimension.
If \code{TRUE}, \code{X}'s dimension is \eqn{n \times m}.}

\item{B}{An optional integer (defaults to 100) of number of bootstrap replications.}

\item{verbose}{A Boolean that controls the returned values. If \code{verbose} is FALSE (default), only the estimated \eqn{\boldsymbol{\Lambda}}
is returned. If TRUE, \code{est_coanc} returns a list consisting of the estimated \eqn{\boldsymbol{\Lambda}},
a list of \eqn{\boldsymbol{\Lambda}}s over iterations, a list of errors over iterations.}

\item{...}{Other options used to control the estimation of coancestry in \code{\link[popkin:popkin]{popkin::popkin()}}. Passed on to \code{\link[popkin:popkin]{popkin::popkin()}}.}
}
\value{
A named list is returned, containing:
\itemize{
\item \code{test_stat1}: A scalar corresponding to the observed test statistics
\item \code{test_stats0}: A length-\eqn{B} vector corresponding to the test statistics
under the null
\item \code{p_value}: A scalar that may be interpreted as the bootstrap p-value
}
}
\description{
\code{test_indep_antepops()} performs a hypothesis testing on whether the assumption of no coancestry
among antecedent populations holds.
}
