% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_amr}
\alias{X_amr}
\title{Genotypes of AMR subset of 1000 Genomes dataset}
\format{
A 10,000 \eqn{\times} 253 matrix of genotypes. Each element is an integer  of 0's, 1's and 2's.
\describe{
Since the number of loci of the AMR subset of 1000 Genomes dataset is too large for a quick analysis,
we created a subset of this dataset by first applying allele frequency filters and LD-pruning to the
AMR dataset. We then randomly selected 10,000 SNPs out of LD-pruned SNP sets. We also re-ordered individuals
according to their pairwise kinship level. This subset is available in \code{data/X_amr.rda}. This data has
353 individuals and 10,000 loci. These data can be reproduced by scripts \verb{data-raw/\{amr.bash,amr.R\}}.
}
}
\usage{
X_amr
}
\description{
Genotypes of AMR subset of 1000 Genomes dataset
}
\keyword{datasets}
