% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{X_hgdp}
\alias{X_hgdp}
\title{Genotypes of HGDP dataset}
\format{
A 10,000 \eqn{\times} 929 matrix of genotypes. Each element is an integer  of 0's, 1's and 2's.
\describe{
we created a subset of this dataset by first applying allele frequency filters and LD-pruning to the
AMR dataset. We then randomly selected 10,000 SNPs out of LD-pruned SNP sets. This subset is available in
\code{data/X_hgdp.rda}. This data has 929 individuals and 10,000 loci. These data can be reproduced by scripts
\verb{data-raw/\{hgdp.bash,hgdp.R\}}.
}
}
\usage{
X_hgdp
}
\description{
Genotypes of HGDP dataset
}
\keyword{datasets}
